requirements.txt
numpy~=1.20.3
networkx~=2.6.2
pandas~=1.3.2
tqdm~=4.62.2
pillow~=8.3.1
six~=1.16.0
scikit-learn~=0.24.2
opencv-python~=4.7.0
pygame~=2.2.0
pyyaml~=6.0
pytest~=7.2.1
pytorch-nlp
wandb~=0.13.10
imageio~=2.26.0
py_trees~=2.2.3